+++ bbp/kernel/string.c
@@ -40,10 +40,10 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #define MAX_INT_STR 20
 
-uint32 str_copy(const char *src, char *dest, uint32 len){
+uint32 str_copy(const char *src, char *dest, uint64 len){
 	char *sp = (char *)src;
 	char *dp = (char *)dest;
-	uint32 s = 0;
+	uint64 s = 0;
 	while (*sp != 0 && len != 0){
 		// Copy till the end of source or destination has been reached
 		*(dp++) = *(sp++);
@@ -53,7 +53,7 @@ uint32 str_copy(const char *src, char *dest, uint32 len){
 	return s;
 }
 
-uint32 int_to_str(int32 val, char *dest, uint32 len){
+uint64 int_to_str(int64 val, char *dest, uint64 len){
 	static char tmp[MAX_INT_STR + 1];
 	mem_set(0, (uint8 *)tmp, MAX_INT_STR + 1);
 	char *b = (char *)tmp + MAX_INT_STR;
@@ -75,9 +75,9 @@ uint32 int_to_str(int32 val, char *dest, uint32 len){
 	return str_copy(b, dest, len);
 }
 
-int32 str_to_int(const char *src){
+int64 str_to_int(const char *src){
 	char *b = (char *)src;
-	int32 val = 0;
+	int64 val = 0;
 	uint8 negative = 0;
 	// Trim leading spaces
 	while (*b == 0x20 || *b == 0x09){ // space || tab
@@ -100,8 +100,8 @@ int32 str_to_int(const char *src){
 	return val;
 }
 
-uint32 str_len(const char *src){
-	uint32 len = 0;
+uint64 str_len(const char *src){
+	uint64 len = 0;
 	while (*(src++) != 0){
 		// Increment till the end of string
 		len ++;
@@ -109,8 +109,8 @@ uint32 str_len(const char *src){
 	return len;
 }
 
-int32 str_char_pos(const char *haystack, const char needle, uint32 offset){
-	int32 idx = 0;
+int64 str_char_pos(const char *haystack, const char needle, uint64 offset){
+	int64 idx = 0;
 	// Move offset
 	while (offset--){
 		if (*(haystack++) == 0){
